chore(warnings): Phase 4 — MSVC /W3 exclusive warnings (#4031). Principle VIII. (#4061)
## Summary

Refs #4031 (tracking issue stays open — the post-Phase-4 zero-warning CI
enforcement step is still outstanding). Fixes all 12 current MSVC `/W3`
warnings in `src/`, re-derived from the actual `check-windows` CI log
for
the latest `main` run (run 28750499293, commit `3880b2eb`) rather than
trusting the RFC's `104ebca6` snapshot — line numbers had drifted but
all
12 items were still present, unresolved by anything that landed since,
and
no new `src/` items appeared.

**Two files here are Tier 1** (`src/models/RadioModel.cpp`,
`src/core/AudioEngine.cpp`) — flagging per @ten9876's own note in the
issue
thread that Phase 4's "Tier 3" label was wrong, same as Phase 1's was.

- **C4267** `size_t`→`int` (`AudioEngine.cpp` ×2, `OpusCodec.cpp`,
  `RADEEngine.cpp` ×2): each is a buffer-length division
(`byteArray.size() / sizeof(T)`) bounded by realistic audio buffer
sizes,
nowhere near `INT_MAX`. Added explicit `static_cast<int>` to document
the
  intentional narrowing.
- **C4996** `getenv` (`SpectralNR.cpp:283`, inside `#ifdef _WIN32` —
explains
why GCC never saw it): replaced `std::getenv("APPDATA")` with
`qgetenv()`,
  Qt's cross-platform wrapper, adjusting the null-check/`std::string`
  construction to `QByteArray`'s API.
- **C4244** `double`→`float` (`SpectralNR.h` ×3): the `gainMax`/`qspp`/
`gainSmooth` getters return `float` but read from `std::atomic<double>`
  members — explicit `static_cast<float>` on each return.
- **C4305** double truncated to float (`SmartMtrWidget.cpp:526`): `0.6`
  literal into `QColor::setAlphaF(float)` — added the `f` suffix.
- **C4804** `bool` used as operand to `>` (`AutomationServer.cpp`,
  `ThemeManager.cpp`, `RadioModel.cpp` ×2): all four are the identical
pattern, `QHash::remove(key) > 0`. Confirmed in the installed Qt6
headers
  that plain `QHash::remove()` returns `bool` (not a count — that's
  `QMultiHash`), so this was genuinely comparing a bool against `0`, but
produces the exact same result as the bool alone. Per @ten9876's
guidance,
this is the "intentionally boolean, no behavior change" bucket, not the
"stop and flag it" bucket — none of the four are latent bugs. Dropped
the
  redundant `> 0`.

No behavior change in any of the 12 fixes.

## Constitution principle honored

Principle VIII — Evidence Over Assertion. The inventory driving this PR
came from reading the actual `check-windows` CI log for current `main`,
not from re-running the RFC's stale line numbers or asserting the fix
was
correct without checking Qt's actual `QHash::remove()` return type.

## Test plan

- [x] Local build passes (`cmake --build build`) — full clean rebuild of
all 9 touched files, zero new warnings (GCC continues to build these
files warning-free, as before — these are MSVC-only diagnostics GCC
      never emitted, so a GCC build can't confirm the fix, only that
      nothing regressed)
- [ ] Behavior verified on a real radio — n/a, no behavior change; see
      note below on MSVC verification
- [x] Existing tests pass (CI) — full local `ctest` 63/63 green,
including
`spectral_nr_test` and `theme_manager_test` which directly exercise
      two of the touched files
- [x] Reproduction steps documented — see Summary

**MSVC verification note:** No MSVC/Wine toolchain available in this dev
environment, so this PR's actual pass/fail signal is a green
`check-windows` CI run once pushed — GCC can't reproduce or confirm
C4xxx
diagnostics at all. Requesting that be checked before merge rather than
relying on the local build above.

## Checklist

- [x] Commits are signed (`docs/COMMIT-SIGNING.md`)
- [x] No new flat-key `AppSettings` calls
- [x] Code is clean-room
- [x] All meter UI uses `MeterSmoother` — n/a, no meter UI touched
- [ ] Documentation updated — n/a, no user-visible behavior changed
- [ ] Security-sensitive changes reference a GHSA — n/a

+++ src/core/AudioEngine.cpp
@@ -5660,7 +5660,7 @@ void AudioEngine::processNr2(const QByteArray& stereoPcm,
 
 QByteArray AudioEngine::applyBoost(const QByteArray& pcm, float gain) const
 {
-    const int nSamples = pcm.size() / sizeof(int16_t);
+    const int nSamples = static_cast<int>(pcm.size() / sizeof(int16_t));
     const auto* src = reinterpret_cast<const int16_t*>(pcm.constData());
     QByteArray out(pcm.size(), Qt::Uninitialized);
     auto* dst = reinterpret_cast<int16_t*>(out.data());
@@ -6783,7 +6783,7 @@ void AudioEngine::feedDaxTxAudio(const QByteArray& inPcm)
     // P/CW mic gauge shows DAX audio level regardless of mic profile (#517)
     {
         const auto* src = reinterpret_cast<const float*>(float32pcm.constData());
-        const int samples = float32pcm.size() / sizeof(float);
+        const int samples = static_cast<int>(float32pcm.size() / sizeof(float));
         float peak = 0.0f;
         double sumSq = 0.0;
         for (int i = 0; i < samples; ++i) {
